Data Loading Logic/requirements.txt
functions-framework==3.*
PyMySQL==1.1.1
cloud-sql-python-connector==1.14.0
google-cloud-storage==2.18.2
pandas==2.2.3
nltk==3.9.1
joblib==1.4.2
matplotlib==3.9.2
scikit-learn==1.5.2
openpyxl==3.1.5
SQLAlchemy==2.0.7
fsspec==2024.10.0
gcsfs==2024.10.0
